ARIA-main/ARIA-main/voice_chat_gemini/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
requests>=2.31.0
